fourth time is the charm?
Finally found a micro that works with the only USB library that has a mute buttongit status! Nico Hood HID Project supports consumer keys, which do apparently work on windows, but I think it's a totally different device descriptor vs a keyboard. Unfortunately, it only works on AVR micros. The adafruit arduino library for usb hid keyboards works with rp2040 but does NOT support consumer keys (volume, media, etc). Adafruit DOES support this in circuit python, so I should probably switch to that. I _could_ figure out how to write custom descriptors and reports for the rp2040 in arduino (or even c++) and that would be cool but NOT a today project. That's a whole day, or maybe a few whole days. So, I finally gave in and used the Arduino Micro I have laying here already on a breadboard. It already had a button, so I got that set up as a mute key. It's very simple, but it works! Later I'll figure out the encoder. But at least I have a proof of concept, so long as I use this specific micro.