it works!
Turns out you need an extra function to convert between pin number and interrupt number on AVR micros b/c not every pin is an interrupt. Also, encoder position should _really_ be returning an int, not a uint. But I'm proud of my hack so.

Next: mute on press. Also, scaling, because 100% is like half a rotation lol.