try button
The button on my encoder is apparently broken. That's a bummer. comment that back out. I don't know if it broke a long time ago and I didn't notice, or if having this heavy wheel on it broke it. But pressing the button doesn't toggle the button pin, and turning the encoder does. I _guess_ it could be a wiring thing? but that seems unlikely, I even wrapped everything in electrical tape. So, for now, no mute.